fixed dropdownoption to store ID properly
