+++ Manatee.Trello.sln
@@ -23,6 +23,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SolutionItems", "SolutionIt
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Manatee.Trello.UnitTests", "Manatee.Trello.UnitTests\Manatee.Trello.UnitTests.csproj", "{7842959C-F2CE-4DDC-9A42-C11EDAB548E6}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Manatee.Trello.CardRepeater", "Manatee.Trello.CardRepeater\Manatee.Trello.CardRepeater.csproj", "{FA232CCF-7967-4119-B398-34E26D1D18CC}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -99,6 +101,22 @@ Global
 		{7842959C-F2CE-4DDC-9A42-C11EDAB548E6}.Release-4.5|Any CPU.Build.0 = Release|Any CPU
 		{7842959C-F2CE-4DDC-9A42-C11EDAB548E6}.Release-Mono|Any CPU.ActiveCfg = Release|Any CPU
 		{7842959C-F2CE-4DDC-9A42-C11EDAB548E6}.Release-Mono|Any CPU.Build.0 = Release|Any CPU
+		{FA232CCF-7967-4119-B398-34E26D1D18CC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FA232CCF-7967-4119-B398-34E26D1D18CC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FA232CCF-7967-4119-B398-34E26D1D18CC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FA232CCF-7967-4119-B398-34E26D1D18CC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FA232CCF-7967-4119-B398-34E26D1D18CC}.Release-3.5|Any CPU.ActiveCfg = Release|Any CPU
+		{FA232CCF-7967-4119-B398-34E26D1D18CC}.Release-3.5|Any CPU.Build.0 = Release|Any CPU
+		{FA232CCF-7967-4119-B398-34E26D1D18CC}.Release-3.5C|Any CPU.ActiveCfg = Release|Any CPU
+		{FA232CCF-7967-4119-B398-34E26D1D18CC}.Release-3.5C|Any CPU.Build.0 = Release|Any CPU
+		{FA232CCF-7967-4119-B398-34E26D1D18CC}.Release-4.0|Any CPU.ActiveCfg = Release|Any CPU
+		{FA232CCF-7967-4119-B398-34E26D1D18CC}.Release-4.0|Any CPU.Build.0 = Release|Any CPU
+		{FA232CCF-7967-4119-B398-34E26D1D18CC}.Release-4.0C|Any CPU.ActiveCfg = Release|Any CPU
+		{FA232CCF-7967-4119-B398-34E26D1D18CC}.Release-4.0C|Any CPU.Build.0 = Release|Any CPU
+		{FA232CCF-7967-4119-B398-34E26D1D18CC}.Release-4.5|Any CPU.ActiveCfg = Release|Any CPU
+		{FA232CCF-7967-4119-B398-34E26D1D18CC}.Release-4.5|Any CPU.Build.0 = Release|Any CPU
+		{FA232CCF-7967-4119-B398-34E26D1D18CC}.Release-Mono|Any CPU.ActiveCfg = Release|Any CPU
+		{FA232CCF-7967-4119-B398-34E26D1D18CC}.Release-Mono|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -107,6 +125,7 @@ Global
 		{598234BA-460F-44C4-A3DB-E5257B572998} = {29C1914C-949C-4B23-BC09-94D0843D0B1B}
 		{F6AA1620-3117-4140-A82B-03BF9C854836} = {29C1914C-949C-4B23-BC09-94D0843D0B1B}
 		{7842959C-F2CE-4DDC-9A42-C11EDAB548E6} = {29C1914C-949C-4B23-BC09-94D0843D0B1B}
+		{FA232CCF-7967-4119-B398-34E26D1D18CC} = {1A311328-1FF7-421E-95D9-05B3984AD683}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {2904F187-70FC-4F80-8DDC-88B4F3FFE7CF}
